refactor: move API routes into router modules
