+++ server/router/analyze.js
@@ -1,0 +1,64 @@
+import express from 'express';
+import OpenAI from 'openai';
+
+const router = express.Router();
+
+const openai = new OpenAI({ apiKey: process.env.OPENAI_API_KEY });
+
+router.post(
+  '/api/analyze-book-image',
+  express.raw({ type: 'multipart/form-data', limit: '10mb' }),
+  async (req, res) => {
+    try {
+      const boundaryMatch = req.headers['content-type']?.match(/boundary=(.*)$/);
+      if (!boundaryMatch) {
+        return res.status(400).json({ error: 'Invalid form data' });
+      }
+      const boundary = Buffer.from(`--${boundaryMatch[1]}`);
+      const body = req.body;
+      let start = body.indexOf(boundary) + boundary.length;
+      start = body.indexOf('\r\n\r\n', start) + 4;
+      const end = body.indexOf(boundary, start);
+      const imageBuffer = body.slice(start, end - 2);
+
+      console.log(`Received ${imageBuffer.length} bytes for analyze-book-image`);
+
+      const imageBase64 = imageBuffer.toString('base64');
+      const messages = [
+        {
+          role: 'user',
+          content: [
+            {
+              type: 'text',
+              text:
+                'Extract the book title, author, description and ISBN from this book cover. ' +
+                'Respond in JSON with keys "title", "author", "description", "isbn".',
+            },
+            {
+              type: 'image_url',
+              image_url: `data:image/jpeg;base64,${imageBase64}`,
+            },
+          ],
+        },
+      ];
+      const chat = await openai.chat.completions.create({
+        model: 'gpt-4-vision-preview',
+        messages,
+        temperature: 0,
+      });
+      let metadata = { title: '', author: '', description: '', isbn: '' };
+      try {
+        metadata = JSON.parse(chat.choices[0].message.content);
+      } catch (e) {
+        console.error('Failed to parse OpenAI response:', chat.choices[0].message.content);
+      }
+
+      res.json(metadata);
+    } catch (err) {
+      console.error('analyze-book-image error', err);
+      res.status(500).json({ error: 'Failed to analyze image' });
+    }
+  }
+);
+
+export default router;
